% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw_multiple_p.R
\name{aipw_Multiple_p}
\alias{aipw_Multiple_p}
\title{apply aipw_single_p for q times}
\usage{
aipw_Multiple_p(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  learners,
  control,
  num_cf,
  n_split
)
}
\arguments{
\item{data}{similar as aipw_single_p() function}

\item{exposure}{similar as aipw_single_p() function}

\item{outcome}{similar as aipw_single_p() function}

\item{covarsT}{similar as aipw_single_p() function}

\item{covarsO}{similar as aipw_single_p() function}

\item{learners}{similar as aipw_single_p() function}

\item{control}{similar as aipw_single_p() function}

\item{num_cf}{number of partitions}

\item{n_split}{similar as aipw_single_p() function}
}
\value{
a tibble of the estimates
}
\description{
apply aipw_single_p for q times
}
\examples{
sum(1:5)


}
