% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw_single_p.R
\name{aipw_single_p}
\alias{aipw_single_p}
\title{Estimate ATE using p splits cross-fit procedure for AIPW estimator}
\usage{
aipw_single_p(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  learners,
  control,
  n_split
)
}
\arguments{
\item{data}{a data frame of tibble}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable}

\item{covarsT}{a vector of names of covaraites for treatment model}

\item{covarsO}{a vector of names of covaraites for outcome model}

\item{learners}{similar as \link{https://rdrr.io/cran/SuperLearner/man/SuperLearner.html#heading-2}(SL.library() in Superlearner::Superlearner())}

\item{control}{similar as \link{https://rdrr.io/cran/SuperLearner/man/SuperLearner.html#heading-2}(cvControl() in Superlearner::Superlearner())}

\item{n_split}{number of splits}
}
\value{
a tibble of estimates
}
\description{
Estimate ATE using p splits cross-fit procedure for AIPW estimator
}
\examples{
sum(1:5)

}
