% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cf-data}
\alias{cf-data}
\alias{cf_scale}
\alias{cf_competition}
\alias{cf_region}
\alias{cf_division}
\title{Key-value lists used as arguments to cf functions}
\format{
An object of class \code{list} of length 3.

An object of class \code{list} of length 7.

An object of class \code{list} of length 7.

An object of class \code{list} of length 36.
}
\usage{
cf_scale

cf_competition

cf_region

cf_division
}
\description{
A set of incomplete but still useful key-value lists that map a query
parameter of interest to its underlying integer or string value used in
the actual request string. For example, the \code{Women} division has a query
parameter value of \code{2L}, but you can use \code{cf_division$Women} anywhere you
see a \code{division} argument.
\itemize{
\item \code{cf_scale}: The scale used in the workout. Includes \code{rx}, \code{scaled}, and
\code{foundations} scales.
\item \code{cf_competition}: The competition to retrieve data for. The two most
important are \code{open} and \code{games}.
\item \code{cf_region}: The region to filter down to. From what I can tell, using
\code{worldwide} as the region will pull in all the data, and then you can filter
from there. This is an incomplete set of data.
\item \code{cf_division}: The division to retrieve data for. Most relevant are \code{Men}
and \code{Women}, but you can also filter for an age group, like \code{Women (50-54)}.
This is an incomplete set of data.
}
}
\examples{
cf_scale
}
\keyword{datasets}
