% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\title{A simulated data set having 2000 rows and 7 columns.}
\format{
\subsection{\code{data}}{

A data frame with 2000 rows and 7 columns:
\describe{
\item{Y}{Atherosclerotic cardiovascular disease (ASCVD) status}
\item{statin}{exposure of statin usage}
\item{age}{patient's age}
\item{ldl_log}{log of low-density lipoprotein measurement}
\item{diabetes}{Diabetes status}
\item{risk_score}{Risk score of ASCVD for each patient}
\item{risk_score_cat}{Risk category of ASCVD for each patient. The patient belongs to 3 category has the greatest risk}

}
}
}
\usage{
data
}
\description{
A simulated data set having 2000 rows and 7 columns.
}
\keyword{datasets}
