% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DC_tmle_g1_k.R
\name{DC_tmle_g1_k}
\alias{DC_tmle_g1_k}
\title{Estimate Average Treatment Effect (ATE) using the TMLE estimator with cross-fit algorithm (generalization 1)}
\usage{
DC_tmle_g1_k(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  family.y = "binomial",
  learners,
  control,
  n_split,
  num_cf,
  rand_split = FALSE,
  gbounds = NULL,
  Qbounds = 5e-04,
  seed = 146,
  conf.level = 0.95,
  stat = "median"
)
}
\arguments{
\item{data}{A data frame or tibble containing the dataset.}

\item{exposure}{Name of the exposure variable (treatment indicator).}

\item{outcome}{Name of the outcome variable.}

\item{covarsT}{A vector of names of covariates for the treatment model.}

\item{covarsO}{A vector of names of covariates for the outcome model.}

\item{family.y}{The family for the outcome model. It can be \code{binomial()} (default) or \code{"gaussian"}.}

\item{learners}{Similar to \code{SL.library()} in the \code{SuperLearner} package, specifying the learners to be used.}

\item{control}{Similar to \code{cvControl()} in the \code{SuperLearner} package, specifying the control parameters.}

\item{n_split}{Number of splits used. The default is \code{n_split = 3}.}

\item{num_cf}{Number of repetitions. The default is 5.}

\item{rand_split}{Logical value. If \code{FALSE} (default), discordant splits for exposure and outcome models are chosen systematically; otherwise, they are chosen randomly.}

\item{gbounds}{A numeric vector of length 2 specifying the bounds for truncation of predicted probabilities of exposure. The defaults are 5/sqrt(n)log(n) and 1-5/sqrt(n)log(n). See \code{tmle::tmle()} for more information.}

\item{Qbounds}{A numeric vector of length 2 specifying the bounds for predicted probabilities of outcomes. The defaults are 5e-04 and 1-5e-04.}

\item{seed}{Numeric value for reproducibility of splits.}

\item{conf.level}{Confidence level for confidence intervals. Default is 0.95.}

\item{stat}{Name of the summary statistics, "median" (default) or "mean", to calculate the overall ATE from the repetition-specific estimates.}
}
\value{
A tibble containing risk (or mean) difference (\code{ATE}), standard error (\code{se}), lower and upper confidence intervals (\code{lower.ci} and \code{upper.ci}, respectively).
}
\description{
This function estimates the Average Treatment Effect (ATE) by applying the Targeted Maximum Likelihood Estimator (TMLE) using a cross-fit approach. The package provides tools for data simulation and model fitting.
}
\section{Data Generation}{

The package includes a function, \code{gen.data}, to generate synthetic datasets for simulation studies. The function creates covariates and treatment indicators, and can generate counterfactual outcomes based on user-defined parameters.
}

\examples{

# Generate a synthetic dataset with 1200 observations
data <- gen.data(n = 1200, do.misp = TRUE, my.transform = TRUE, seed = 234)
head(data)

# Fit the TMLE model to the generated data
fit_tmle_g1 <- DC_tmle_g1_k(data = data, exposure = "X", outcome = "Y",
                            covarsT = c("C1", "C2", "C3", "C4"),
                            covarsO = c("C1", "C2", "C3", "C4"),
                            family.y = "gaussian",
                            learners = c("SL.glm", "SL.mean"),
                            control = list(V = 3, stratifyCV = FALSE, shuffle = TRUE, validRows = NULL),
                            num_cf = 5, n_split = 3, rand_split = FALSE,
                            Qbounds = 0.0025, gbounds = 0.0025, seed = 236,
                            conf.level = 0.95, stat = "median")

# Display the results
print(fit_tmle_g1)

}
\references{
Kang, J. D. Y., & Schafer, J. L. (2007). Demystifying Double Robustness: A Comparison of Alternative Strategies for Estimating a Population Mean from Incomplete Data. \emph{Statistical Science}, 22(4), 523-539.
}
