% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{dml}
\alias{dml}
\alias{dml_fast}
\title{Double Machine Learning Estimates}
\usage{
dml(f, d, psi, psi_grad, psi_op, n = 101, nw = 4, dml_seed = NULL,
  ...)

dml_fast(f, d, psi, psi_grad, psi_op, dml_seed = NULL)
}
\arguments{
\item{f}{an object of class formula representing the model to be fitted.}

\item{d}{a dataframe containing the variables in f.}

\item{psi}{function that gives the value of the Neyman-Orthogonal moment at a
given value of theta}

\item{psi_grad}{function that gives the gradient of psi with respect to theta}

\item{psi_plr_op}{function that gives the variance estimator at a given
value of theta.}
}
\value{
\code{dml} returns an object of class "dml" with the following components:
\describe{
   \item{coefficients}{a named vector of coefficients.}
   \item{vcov}{variance-covariance matrix of the main parameters.}
   \item{nobs}{number of observations used}
}
}
\description{
Double Machine Learning Estimates
}
\examples{
library("MASS")
data(Boston)

crime_dml <-
  "crim ~ zn + indus + nox" \%>\%
  as.formula() \%>\%
  dml(Boston, psi_plr, psi_plr_grad, psi_plr_op, n = 101)

}
\references{
V. Chernozhukov, D. Chetverikov, M. Demirer, E. Duflo, C. Hansen,
W. Newey, and J. Robins. Double/debiased machine learning for treatment and
structural parameters.The Econometrics Journal, 21(1):C1–C68, 2018a.
}
