% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generlization_code.R
\name{generate_placeholder_output}
\alias{generate_placeholder_output}
\title{Title Additional code to generalize the results}
\usage{
generate_placeholder_output(SL.library, n_split)
}
\arguments{
\item{SL.library}{the learners we input}

\item{n_split}{number of splits}
}
\value{
it return more general results
}
\description{
Title Additional code to generalize the results
}
\examples{

# See the README file

}
