% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ObsData.R
\docType{data}
\name{ObsData}
\alias{ObsData}
\title{Description of ObsData}
\format{
A data frame with 1200 rows and 6 columns.
}
\usage{
ObsData
}
\description{
This dataset contains information about 1200 patients (simulated) on outcome (Y), exposure (X), confounders (C1, C2, C3, and C4). The simulation details are described in Kang et al. (2007).
}
\keyword{datasets}
