% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_aipw.R
\name{par_aipw}
\alias{par_aipw}
\title{Estimate Average Treatment Effect (ATE) using from AIPW estimator using cross-fit algorithm with parallelizing}
\usage{
par_aipw(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  family.y = binomial(),
  learners = c("SL.glm", "SL.glmnet", "SL.xgboost"),
  control,
  num_cf = 5,
  n_split = 3,
  rand_split = FALSE,
  gbound = 0.025,
  alpha = 1e-17,
  num_core = detectCores(),
  seed = 145
)
}
\arguments{
\item{data}{a data frame of tibble}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable}

\item{covarsT}{a vector of names of covaraites for treatment model}

\item{covarsO}{a vector of names of covaraites for outcome model}

\item{family.y}{it is the family for outcome model. It can \code{binomial() (default)} or \code{"gaussian"}}

\item{learners}{similar as \code{SL.library()} in \code{SuperLearner} package.}

\item{control}{similar as \code{cvControl()} in \code{SuperLearner} package.}

\item{num_cf}{number of repetition done. The default is 5.}

\item{n_split}{number of splits used, default \code{n_split = 3}}

\item{rand_split}{logical value; if be FALSE \code{(default)}, discordant splits for exposure and outcome model are chosen systematically; otherwise chosen randomly.}

\item{gbound}{value between (0,1) for truncation of predicted probabilities. The defaults are 0.025 and 0.975. See \code{tmle::tmle()} for more information.}

\item{alpha}{used to keep predicted initial values bounded away from (0,1) for logistic fluctuation. The defaults are 1e-17 and 1-1e-17.}

\item{num_core}{an integer indicating the number of cores to be used for parallelization. The default is all the cores that user owns in the system.}

\item{seed}{numeric value to reproduce the splits distribution}
}
\value{
a tibble of the estimates
}
\description{
Estimate Average Treatment Effect (ATE) using from AIPW estimator using cross-fit algorithm with parallelizing
}
\examples{

# See the README file for details

sum(1:5)


}
