% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moments.R
\name{psi_plr}
\alias{psi_plr}
\title{Double Machine Learning Moments}
\usage{
psi_plr(theta, Y, D, gamma, delta)
}
\arguments{
\item{theta}{scalar or vector of the treatment effect parameter(s).}

\item{Y}{a vector of the outcome variable.}

\item{D}{a vector or data.frame of the treatment variable(s) of interest.}

\item{gamma}{a vector of E[Y|X=x], where X is a vector of controls.}

\item{delta}{a vector or data.frame of E[D|X=x].}
}
\description{
Moment functions for partial linear model and partial linear poisson model.
}
