% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_functions.R
\name{regression_forest2}
\alias{regression_forest2}
\title{Machine Learning Functions}
\usage{
regression_forest2(f, d, num.trees = 1000, ...)
}
\arguments{
\item{f}{an object of class formula representing the model to be fitted.}

\item{d}{a dataframe containing the variables in f.}

\item{...}{additional arguments taht can be passed to
\code{\link[grf:regression_forest]{regression_forest}},such as num.trees,
honesty,honesty.fraction,seed.}
}
\description{
\code{regression_forest2} estimates a regression forest with a formula
  instead of manually passing matrices.
\code{predict_rf2} predicts based on the forest ouputted from
  \code{regression_forest2}.
}
