% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statin_sim_data.R
\docType{data}
\name{statin_sim_data}
\alias{statin_sim_data}
\title{A simulated data set having 3000 rows and 7 columns.}
\format{
\subsection{\code{statin_sim_data}}{

A data frame with 3000 rows and 7 following columns:
\describe{
\item{Y}{Atherosclerotic cardiovascular disease (ASCVD) status}
\item{statin}{exposure of statin usage}
\item{age}{patient's age}
\item{ldl_log}{log of low-density lipoprotein measurement}
\item{diabetes}{Diabetes status}
\item{risk_score}{Risk score of ASCVD for each patient}
\item{risk_score_cat}{Risk category of ASCVD for each patient. The patient belongs to 3 category has the greatest risk}

}
}
}
\usage{
statin_sim_data
}
\description{
The \code{statin_sim_data} is generated from the guidelines described by \verb{Zivich and Breskin (2021)}
}
\references{
\enumerate{
\item Paul N Zivich and Alexander Breskin (2021),
Machine learning for causal inference: on the use of cross-fit estimators.
\emph{Epidemiology}, 32(3):393–401.
\doi{10.1097/EDE.0000000000001332}
}
}
\keyword{datasets}
