% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_single_g1_p.R
\name{tmle_single_g1_p}
\alias{tmle_single_g1_p}
\title{Estimate Average Treatment Effect (ATE) using TMLE estimator using cross-fit algorithm for single repetition (generalization 1)}
\usage{
tmle_single_g1_p(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  family.y,
  learners,
  control,
  n_split,
  rand_split,
  gbounds,
  Qbounds,
  seed = 146
)
}
\arguments{
\item{data}{a data frame of tibble}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable}

\item{covarsT}{a vector of names of covaraites for treatment model}

\item{covarsO}{a vector of names of covaraites for outcome model}

\item{family.y}{it is the family for outcome model. It can \code{"binomial" (default)} or \code{"gaussian"}}

\item{learners}{similar as \code{SL.library()} in \code{SuperLearner} package.}

\item{control}{similar as  \code{cvControl()} in \code{SuperLearner} package.}

\item{n_split}{number of splits used, default \code{n_split = 3}}

\item{rand_split}{logical value; if be FALSE \code{(default)}, discordant splits for exposure and outcome model are chosen systematically; otherwise chosen randomly.}

\item{gbounds}{value between (0,1) for truncation of predicted probabilities of exposure. The defaults are 5/sqrt(n)log(n) and 1-5/sqrt(n)log(n). See \code{tmle::tmle()} for more information.}

\item{Qbounds}{used to keep predicted probabilities of outcomes values bounded away from (0,1). The defaults are 5e-04 and 1-5e-04.}

\item{seed}{numeric value to reproduce the splits distribution}
}
\value{
A tibble containing risk (or mean) difference (\code{rd}), variance of the estimated \code{rd}.
}
\description{
Estimate Average Treatment Effect (ATE) using TMLE estimator using cross-fit algorithm for single repetition (generalization 1)
}
\examples{

# See the README file for details

sum(1:5)




}
