% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_single_p.R
\name{tmle_single_p}
\alias{tmle_single_p}
\title{Estimate ATE using TMLE estimator using cross-fit algorithm (single repetition)}
\usage{
tmle_single_p(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  family.y,
  learners,
  control,
  n_split,
  rand_split,
  seed
)
}
\arguments{
\item{data}{a data frame of tibble}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable}

\item{covarsT}{a vector of names of covaraites for treatment model}

\item{covarsO}{a vector of names of covaraites for outcome model}

\item{family.y}{it is the family for outcome model. It can \code{binomial() (default)} or \code{"gaussian"}}

\item{learners}{similar as \code{\link[Superlearner:SL.library()]{Superlearner::SL.library()}}}

\item{control}{similar as  \code{\link[Superlearner:cvControl()]{Superlearner::cvControl()}}}

\item{n_split}{number of splits used, default \code{n_split = 3}}

\item{rand_split}{logical value; if be FALSE \code{(default)}, discordant splits for exposure and outcome model are chosen systematically; otherwise chosen randomly.}

\item{seed}{numeric value to reproduce the splits distribution}
}
\value{
a tibble of estimates.
}
\description{
Estimate ATE using TMLE estimator using cross-fit algorithm (single repetition)
}
\examples{

# See the README file for details

sum(1:5)




}
