% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw_multiple_p.R
\name{aipw_multiple_p}
\alias{aipw_multiple_p}
\title{apply aipw_single_p for num_cf times}
\usage{
aipw_multiple_p(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  learners,
  control,
  num_cf,
  n_split,
  rand_split = TRUE,
  seed = 145
)
}
\arguments{
\item{data}{similar as aipw_single_p() function}

\item{exposure}{similar as aipw_single_p() function}

\item{outcome}{similar as aipw_single_p() function}

\item{covarsT}{similar as aipw_single_p() function}

\item{covarsO}{similar as aipw_single_p() function}

\item{learners}{similar as aipw_single_p() function}

\item{control}{similar as aipw_single_p() function}

\item{num_cf}{number of partitions}

\item{n_split}{similar as aipw_single_p() function}

\item{rand_split}{logical value; if be TRUE, discordant splits for exposure and outcome model are chosen at random ; otherwise chosen systematically.}

\item{seed}{numeric value to reproduce the splits}
}
\value{
a tibble of the estimates
}
\description{
apply aipw_single_p for num_cf times
}
\examples{
sum(1:4)


}
