% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aipw_single_p.R
\name{aipw_single_p}
\alias{aipw_single_p}
\title{Estimate Average Treatment Effect (ATE) using AIPW estimator with cross-fit algorithm (single repetition)}
\usage{
aipw_single_p(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  family.y = "binomial",
  learners,
  control,
  n_split = 5,
  rand_split = FALSE,
  gbound = 0.025,
  alpha = 1e-17,
  seed = 146
)
}
\arguments{
\item{data}{a data frame of tibble}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable}

\item{covarsT}{a vector of names of covaraites for treatment model}

\item{covarsO}{a vector of names of covaraites for outcome model}

\item{family.y}{it is the family for outcome model. It can \code{binomial() (default)} or \code{"gaussian"}}

\item{learners}{similar as \code{SL.library()} in \code{SuperLearner} package.}

\item{control}{similar as  \code{cvControl()} in \code{SuperLearner} package.}

\item{n_split}{number of splits used, default \code{n_split = 3}}

\item{rand_split}{logical value; if be FALSE \code{(default)}, discordant splits for exposure and outcome model are chosen systematically; otherwise chosen randomly.}

\item{gbound}{value between (0,1) for truncation of predicted probabilities. See \code{tmle::tmle()} for more information.}

\item{alpha}{used to keep predicted initial values bounded away from (0,1) for logistic fluctuation.}

\item{seed}{numeric value to reproduce the splits distribution}
}
\value{
a tibble of estimates.
}
\description{
Estimate Average Treatment Effect (ATE) using AIPW estimator with cross-fit algorithm (single repetition)
}
\examples{

# See the README file for details


sum(1:5)

}
