% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfit-package.R
\docType{package}
\name{crossfit-package}
\alias{crossfit}
\alias{crossfit-package}
\title{crossfit: CrossFit API}
\description{
Provides tooling for accessing the CrossFit API. This includes data about the CrossFit Open and the CrossFit Games.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/crossfit}
  \item \url{https://davisvaughan.github.io/crossfit/}
  \item Report bugs at \url{https://github.com/DavisVaughan/crossfit/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{dvaughan1130@gmail.com}

}
\keyword{internal}
