% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{dml}
\alias{dml}
\title{Double Machine Learning Estimates}
\usage{
dml(f, d, psi, psi_grad, psi_op, n = 101, nw = 4, dml_seed = NULL,
  ml, poly_degree = 3, drop_na = FALSE, ...)
}
\arguments{
\item{f}{an object of class formula representing the model to be fitted.}

\item{d}{a dataframe containing the variables in f.}

\item{psi}{function that gives the value of the Neyman-Orthogonal moment at a
given value of theta}

\item{psi_grad}{function that gives the gradient of psi with respect to theta}

\item{ml}{}

\item{psi_plr_op}{function that gives the variance estimator at a given
value of theta.}
}
\value{
\code{dml} returns an object of class "dml" with the following components:
\describe{
   \item{coefficients}{a named vector of coefficients.}
   \item{vcov}{variance-covariance matrix of the main parameters.}
   \item{nobs}{number of observations used}
   \item{call}{original function call with given arguments}
}
}
\description{
Double Machine Learning Estimates
}
\examples{
# Effect of temperature and precipitation on corn yield in the presence of
# time and locational effects
data(corn_yield)

yield_dml <-
  "logcornyield ~ lower + higher + prec_lo + prec_hi | year + fips" \%>\%
  as.formula() \%>\%
  dml(corn_yield, psi_plr, psi_plr_grad, psi_plr_op, n = 5,
  ml = "regression_forest", dml_seed = 123)

}
\references{
V. Chernozhukov, D. Chetverikov, M. Demirer, E. Duflo, C. Hansen,
W. Newey, and J. Robins. Double/debiased machine learning for treatment and
structural parameters.The Econometrics Journal, 21(1):C1–C68, 2018a.
}
