% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dml.R
\name{dml}
\alias{dml}
\title{Double Machine Learning Estimates}
\usage{
dml(
  f,
  d,
  model = "linear",
  ml = "lasso",
  n = 101,
  nfolds = 5,
  score = "concentrate",
  workers = 1,
  drop_na = FALSE,
  family = NULL,
  poly_degree = 1,
  ...
)
}
\arguments{
\item{f}{an object of class formula representing the model to be fitted.}

\item{d}{a dataframe containing the variables in f.}

\item{ml}{Machine learning method to be used for estimating nuisance parameters.
Currently it takes in `lasso`, and `rf` for regression forest. Note `rf`
is not available for `score = "finite"`. Default is `ml = "lasso"`.}

\item{n}{Number of times to repeat the sample splitting and take median of
results over the n samples. Default is `n = 100`.}

\item{nfolds}{Number of folds for cross-fitting}

\item{score}{Takes either value `finite` or `concentrate`. `finite` refers to
using the finite nuisance parameter orthogonal score construction, and
`concentrate` refers to using the concentrating out approach.
Default is `score = "finite"`}

\item{workers}{Number of workers to use in running the n dml calculations in
parallel. Default is `workers = 1`, in which case the process is sequential.}

\item{drop_na}{if `TRUE`, then any row with an `NA` value is dropped. Default
is `false`}

\item{family}{if `ml = lasso`, this is passed onto `cv.glmnet` to describe
the response variable type.}

\item{poly_degree}{degree of polynomial for the nuisance parameters.
Default is `poly_degree = 1`.}
}
\value{
\code{dml} returns an object of class "dml" with the following components:
\describe{
   \item{coefficients}{a named vector of coefficients.}
   \item{vcov}{variance-covariance matrix of the main parameters.}
   \item{nobs}{number of observations used}
   \item{call}{original function call with given arguments}
}
}
\description{
Double Machine Learning Estimates
}
\examples{
# Effect of temperature and precipitation on corn yield in the presence of
# time and locational effects
data(corn_yield)
library(magrittr)

yield_dml <-
  "logcornyield ~ lower + higher + prec_lo + prec_hi | year + fips" \%>\%
  as.formula() \%>\%
  dml(corn_yield, "linear", n = 5,
  ml = "regression_forest")

yield_dml_lasso <-
  "logcornyield ~ lower + higher + prec_lo + prec_hi | year + fips" \%>\%
  as.formula() \%>\%
  dml(corn_yield, "linear", n = 5,
  ml = "lasso", family = "gaussian")

}
\references{
V. Chernozhukov, D. Chetverikov, M. Demirer, E. Duflo, C. Hansen,
W. Newey, and J. Robins. Double/debiased machine learning for treatment and
structural parameters.The Econometrics Journal, 21(1):C1–C68, 2018a.
}
