% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{fun_registry_id}
\alias{fun_registry_id}
\title{Internal: assign an integer id to a function signature}
\usage{
fun_registry_id(reg, sig)
}
\arguments{
\item{reg}{A registry environment created by
\code{\link{fun_registry_new}}.}

\item{sig}{A character string code signature.}
}
\value{
An integer id corresponding to \code{sig} within
\code{reg}.
}
\description{
Maps a function code signature (as produced by
\code{\link{fun_code_sig}}) to a small integer id in a registry. If
the signature is new, it is appended to the registry; otherwise, the
existing id is returned. This is used to keep structural signatures
compact.
}
\keyword{internal}
