% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmle_single_p.R
\name{tmle_single_p}
\alias{tmle_single_p}
\title{Estimate ATE using cross-fit procedure for TMLE estimator}
\usage{
tmle_single_p(
  data,
  exposure,
  outcome,
  covarsT,
  covarsO,
  learners,
  control,
  n_split,
  rand_split = TRUE,
  seed = 145
)
}
\arguments{
\item{data}{a data frame of tibble}

\item{exposure}{name of exposure variable}

\item{outcome}{name of outcome variable}

\item{covarsT}{a vector of names of covaraites for treatment model}

\item{covarsO}{a vector of names of covaraites for outcome model}

\item{learners}{similar as \code{\link[Superlearner:SL.library()]{Superlearner::SL.library()}}}

\item{control}{similar as  \code{\link[Superlearner:cvControl()]{Superlearner::cvControl()}}}

\item{n_split}{number of splits}

\item{rand_split}{logical value; if be TRUE, discordant splits for exposure and outcome model are chosen at random ; otherwise chosen systematically.}

\item{seed}{numeric value to reproduce the splits}
}
\value{
a tibble of estimates
}
\description{
Estimate ATE using cross-fit procedure for TMLE estimator
}
\examples{
sum(1:5)


}
